각각 권한을줘서 볼수있는 화면이 제한되도록 만듦, 홈화면으로 시작화면을 구현하려고했으나 30퍼정도 됨

+++ src/main/java/kh/yhjsemi/board/controller/SemiBoardDeleteController.java
@@ -18,31 +18,36 @@ import kh.yhjsemi.student.service.SemiStudentService;
 public class SemiBoardDeleteController extends HttpServlet {
 	private static final long serialVersionUID = 1L;
 
-	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	protected void doPost(HttpServletRequest request, HttpServletResponse response)
+			throws ServletException, IOException {
 		SemiMemberVo loginVo = (SemiMemberVo) request.getSession().getAttribute("loginVo");
 		String bnoStr = request.getParameter("bno");
 		System.out.println(bnoStr);
 		int bno = 0;
-		int result =0;
+		int result = 0;
 		SemiBoardService service = new SemiBoardService();
-		if(loginVo.getMtype().equals("A")) {
-		try {
-			bno = Integer.parseInt(bnoStr);
-			result = service.deleteboard(bno);
-		} catch (Exception e) {
+		if (loginVo != null) {
+			if (loginVo.getMtype().equals("A")) {
+				try {
+					bno = Integer.parseInt(bnoStr);
+					result = service.deleteboard(bno);
+				} catch (Exception e) {
+				}
+				if (result < 1) {
+					request.getSession().setAttribute("msg", "글 삭제 실패!");
+					System.out.println(request.getSession().getAttribute("msg"));
+				} else {
+					request.getSession().setAttribute("msg", "글 삭제 되었습니다");
+					System.out.println(request.getSession().getAttribute("msg"));
+				}
+				response.sendRedirect(request.getContextPath() + "/sm/board/list");
+			} else {
+				request.getSession().setAttribute("msg", "권한이 없습니다");
+				System.out.println(request.getSession().getAttribute("msg"));
+				response.sendRedirect(request.getContextPath() + "/sm/board/list");
+			}
+		} else {
+			response.sendRedirect(request.getContextPath() + "/main");
 		}
-		if(result <1) {
-			request.getSession().setAttribute("msg", "글 삭제 실패!");
-			System.out.println(request.getSession().getAttribute("msg"));
-		}else {
-			request.getSession().setAttribute("msg", "글 삭제 되었습니다");
-			System.out.println(request.getSession().getAttribute("msg"));
-		}
-		response.sendRedirect(request.getContextPath()+"/sm/board/list");
-	}else {
-		request.getSession().setAttribute("msg", "권한이 없습니다");
-		System.out.println(request.getSession().getAttribute("msg"));
-		response.sendRedirect(request.getContextPath()+"/sm/board/list");
 	}
-	}
-	}
+}
